Handle Unicode cell content for width calcs
Assuming all characters display as one cell horizontally, it's
sufficient to replace len() [number of bytes used for the UTF-8 encoded
form of the strings] with unicode/utf8.RuneCountInString().

In fact, there are some code-points defined as wide characters.  We
ignore that for now, to get the majority cases working.