Fix title widths to be UTF-8 aware
We used (*Cell).Width() correctly for all cells, but the title was
special-cased and the width was being measured in bytes, not
characters/runes.

Fix the width of the title to be measured in runes.  Which leaves
us with the same known issue about widths assuming constant-width
runes, not handling combining characters or double-width characters, but
it's much closer to accurate.

Test that the table widths are calculated correctly.

Clean up the table tests a little with a new helper function, reducing
the size of the per-test boilerplate.