Adjust title-handling width calculations
Instead of tracking a table.minWidth, let the title generation cache a
copy of the title `*Cell` into the table structure, so we can
interrogate its length.

Then, entirely within style handling, correctly account for table styles
of borders and padding in determining minimum width for the table to
handle the title, and adjust the rightmost column's record of maximum
cell width to make it longer to account for this.