add skipped tests, highlight grapheme cluster issue
"How long is a string" is not easy in Unicode.  I punted on this when I
added (*Cell).Width(), which just measures rune count.  This was a
better approximation which let us work with a reasonable chunk of
Unicode, but is limited.

Add two tests, each skipped because we currently fail them, but which
highlight the grapheme cluster length problem, with the classic noel vs
noël vs noël example and a fullwidth example.

This should serve as a baseline for anyone who looks to improve matters:
get this test working.

Use `go test -v` to see the skipped tests and get nudged.