render: Fix bug to rerender tables; switch to byte buffers.
The elements of the table were being modified on each call
to Render, leading to headers and titles being compounded.
This fixes the bug by utilizing a deep-ish copy of the table.

In addition, I noticed raw string concatenation in Render
which I swapped out for bytes.Buffer calls which led to a
95% reduction in the duration of the call.

Fixes github.com/apcera/termtables/issues/28